+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @grid-table/basic
 
+## 0.6.29
+
+### Patch Changes
+
+- @grid-table/core@0.6.29
+
 ## 0.6.28
 
 ### Patch Changes